move set_clusterer to contrib
