+++ sktime/contrib/set_clusterer.py
@@ -1,0 +1,46 @@
+# -*- coding: utf-8 -*-
+"""Set classifier function."""
+
+
+from sktime.clustering import TimeSeriesKMeans, TimeSeriesKMedoids
+
+
+def set_clusterer(cls, resample_id=None):
+    """Construct a clusterer.
+
+    Basic way of creating the clusterer to build using the default settings. This
+    set up is to help with batch jobs for multiple problems to facilitate easy
+    reproducability through run_clustering_experiment. You can set up bespoke
+    clusterers and pass them to run_clustering_experiment if you prefer. It also
+    serves to illustrate the base clusterer parameters
+
+    Parameters
+    ----------
+    cls : str
+        indicating which clusterer you want
+    resample_id : int or None, default = None
+        clusterer random seed
+
+    Return
+    ------
+    A clusterer.
+    """
+    name = cls.lower()
+    # Distance based
+    if name == "kmeans" or name == "k-means":
+        return TimeSeriesKMeans(
+            n_clusters=5,
+            max_iter=50,
+            averaging_algorithm="mean",
+            random_state=resample_id,
+        )
+    if name == "kmedoids" or name == "k-medoids":
+        return TimeSeriesKMedoids(
+            n_clusters=5,
+            max_iter=50,
+            averaging_algorithm="mean",
+            random_state=resample_id,
+        )
+
+    else:
+        raise Exception("UNKNOWN CLUSTERER")
